# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JARVIS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/executions": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "List Executions",
        "operationId": "list_executions_v1_executions_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "awaiting_input",
                    "succeeded",
                    "failed",
                    "cancelled",
                    "timed_out"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/executions/{run_id}": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "Get Execution",
        "operationId": "get_execution_v1_executions__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/executions/{run_id}/llm-trace": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "Get Llm Trace",
        "description": "The web-readable half of the JARVIS_LLM_TRACE debug trace (ADR 0014):\nthe process-local buffer holding the actual model request/response per\ncall. Empty entries is a 200, never an error \u2014 flag off, backend\nrestarted since the run, or the run executed in a separate worker\nprocess (distributed mode keeps the log as its only trace).",
        "operationId": "get_llm_trace_v1_executions__run_id__llm_trace_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmTraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/executions/{run_id}/cancel": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Cancel Run",
        "description": "Idempotent. A run live in THIS process gets its runtime token; a\nqueued or foreign-worker run gets a cross-process cancel request\n(ADR 0008 \u00a76) that the owning worker's heartbeat pops. A paused run\n(S10) is finished directly \u2014 no heartbeat holds it \u2014 on the same\nappend-terminal-and-finish path the reaper uses, so cancelling an\nawaiting_input run is immediate, not cooperative. A finished run is a\nno-op that reports its current status.",
        "operationId": "cancel_run_v1_executions__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/executions/{run_id}/resume": {
      "post": {
        "tags": [
          "executions"
        ],
        "summary": "Resume Run",
        "description": "Human-in-the-loop resume (S10, ADR 0010 \u00a74): merge the answer into\nthe paused run's queue payload and block until the resumed segment ends\n(like /run). A segment can pause again \u2014 the route then returns the\nstill-awaiting row and the client answers the next question.",
        "operationId": "resume_run_v1_executions__run_id__resume_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/executions/{run_id}/events": {
      "get": {
        "tags": [
          "executions"
        ],
        "summary": "List Events",
        "operationId": "list_events_v1_executions__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "cancelled",
          "status"
        ],
        "title": "CancelResult"
      },
...
      "EventList": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CursorEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "events"
        ],
        "title": "EventList"
      },
      "ExecutionDetail": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunResult"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          },
          "tool_executions": {
            "items": {
              "$ref": "#/components/schemas/ToolResult"
            },
            "type": "array",
            "title": "Tool Executions"
          },
          "names": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "messages",
          "tool_executions"
        ],
        "title": "ExecutionDetail"
      },
      "ExecutionList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunResult"
            },
            "type": "array",
            "title": "Items"
          },
          "names": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ExecutionList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImagePart": {
        "properties": {
          "type": {
            "type": "string",
            "const": "image",
            "title": "Type",
            "default": "image"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ImagePart",
        "description": "Reserved for multimodal phases; not produced by Phase 1 code."
      },
...
      "LlmTraceEntry": {
        "properties": {
          "iteration": {
            "type": "integer",
            "title": "Iteration"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "at": {
            "type": "string",
            "title": "At"
          },
          "request": {
            "additionalProperties": true,
            "type": "object",
            "title": "Request"
          },
          "response": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "iteration",
          "method",
          "provider",
          "model",
          "at",
          "request"
        ],
        "title": "LlmTraceEntry",
        "description": "One model call's actual request/response (ADR 0014, debug only) \u2014\nserved from the process-local buffer, never persisted."
      },
      "LlmTraceResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/LlmTraceEntry"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "entries"
        ],
        "title": "LlmTraceResponse"
      },
...
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "system",
              "developer",
              "user",
              "assistant",
              "tool"
            ],
            "title": "Role"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TextPart"
                    },
                    {
                      "$ref": "#/components/schemas/ImagePart"
                    }
                  ]
                },
                "type": "array"
              }
            ],
            "title": "Content",
            "default": ""
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ToolCall"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role"
        ],
        "title": "Message"
      },
...
      "ResumeBody": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "tool_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Approval"
          },
          "decisions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decisions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ResumeBody",
        "description": "Body for POST /executions/{id}/resume (S10, ADR 0010 \u00a74; ADR 0011):\nan ask_human pause answers with `content`; a tool-approval pause answers\nwith `tool_approval` (batch shorthand) or `decisions` (per-call\nverdicts \u2014 calls absent from the map are declined, ADR 0011). Exactly\none is present."
      },
...
      "RunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_input",
              "succeeded",
              "failed",
              "cancelled",
              "timed_out"
            ],
            "title": "Status"
          },
          "input": {
            "type": "string",
            "title": "Input",
            "default": ""
          },
          "agent_version_id": {
            "type": "string",
            "title": "Agent Version Id",
            "default": ""
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "final_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Message"
          },
          "total_usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "event_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Cursor"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "agent_id",
          "status"
        ],
        "title": "RunResult"
      },
...
      "TextPart": {
        "properties": {
          "type": {
            "type": "string",
            "const": "text",
            "title": "Type",
            "default": "text"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TextPart"
      },
...
      "ToolCall": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ToolCall"
      },
...
      "ToolResult": {
        "properties": {
          "tool_call_id": {
            "type": "string",
            "title": "Tool Call Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "output": {
            "type": "string",
            "title": "Output"
          },
          "is_error": {
            "type": "boolean",
            "title": "Is Error",
            "default": false
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool_call_id",
          "tool_name",
          "output"
        ],
        "title": "ToolResult"
      },
      "Usage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "extra": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Extra"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Usage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}